+++ Hito-Situacion-Mundial/jwt-example-desafioLatam 2/public/covid19/assets/utils/login.js
@@ -1,0 +1,19 @@
+/*
+//token
+const postData = async (email,password) => {
+    try{
+        const response = await fetch ("http://localhost:3000/api/login", 
+        {
+            method: 'POST',
+            body: JSON.stringify({email: email, password: password}),
+        });
+        const { token } = await response.json();
+        console.log(token)
+        localStorage.setItem('jwt-token', token); 
+        //return token;
+
+    } catch (error){
+        console.error(`Error: ${error}`);
+    }
+};
+*/
