+++ Hito-Situacion-Mundial/jwt-example-desafioLatam 2/public/covid19/index.html
@@ -86,7 +86,7 @@
     <!--ChartJS-->
     <script src="https://cdn.jsdelivr.net/npm/chart.js@3.4.1/dist/chart.min.js"></script>
     <!--Javascript del proyecto-->
-    <script src="./assets/scripts.js"></script>
+    <script type="module" src="./assets/scripts.js"></script>
 
 </body>
